+++ programs/solana/programs/ics07-tendermint/src/test_helpers.rs
@@ -1,4 +1,10 @@
+use std::sync::LazyLock;
+
+use mollusk_svm::result::Check;
+
 pub const PROGRAM_BINARY_PATH: &str = "../../target/deploy/ics07_tendermint";
+
+pub static SUCESS_CHECK: LazyLock<Vec<Check>> = LazyLock::new(|| vec![Check::success()]);
 
 pub mod fixtures {
     use crate::types::{ClientState, ConsensusState, IbcHeight};
